+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './types'
 export * from './TrueSheet'
+export * from './TrueSheet.types'
 export * from './TrueSheetGrabber'
